# update.sh: further speed up `brew update`.