@UniqMartin 2016-08-09
I think this change is basically sound, but it is also pretty invasive. Can you elaborate a little bit on what part of it is responsible for the massive speedup and why?