@MikeMcQuaid 2016-08-10
I've addressed your commits and made it faster still for the `--preinstall` case as a result.

> Can you elaborate a little bit on what part of it is responsible for the massive speedup and why?

Most of it 😀. In short like most dramatic speed-up: do less where possible. Partially avoid recalculating the same values in multiple places but the big speedup: don't try to `status`/`stash`/`merge`/`rebase` when we know nothing has changed (i.e. calling the `merge_or_rebase` method [formerly `pull`]). These are all (relatively) expensive no-ops as they operate on the filesystem data rather than just with variables and filesystem metadata. We've already removed the unnecessary expensive `git fetch` and now `curl` operations too.