@MikeMcQuaid 2016-08-10
CC @ilovezfs who will probably be happy with this speedup and was encouraging me to 🚢 this even before it was this fast.