@UniqMartin 2016-08-10
> I've addressed your commits and made it faster still for the `--preinstall` case as a result.

Nice work! I left a few relatively minor comments, but I really like how this has evolved.

(I have a few ideas for how some things could be implemented more neatly, but that's stuff for a follow-up PR rather than a reason to block this. Hopefully I'll find the time to work on this.)